Move some more tests from test_model to test_fields.
